Refactor GUI/Window utils out of gfx library
