+++ examples/gpu-printing/main.cpp
@@ -7,7 +7,7 @@ using Slang::ComPtr;
 
 #include "gfx/render.h"
 #include "gfx/d3d11/render-d3d11.h"
-#include "gfx/window.h"
+#include "tools/graphics-app-framework/window.h"
 using namespace gfx;
 
 #include <string>
