+++ tools/gfx/d3d/d3d-util.h
@@ -3,10 +3,10 @@
 
 #include <stdint.h>
 
-#include "../../slang-com-helper.h"
+#include "slang-com-helper.h"
 
-#include "../../slang-com-ptr.h"
-#include "../../source/core/slang-list.h"
+#include "slang-com-ptr.h"
+#include "core/slang-list.h"
 
 #include "../flag-combiner.h"
