+++ tools/gfx/vulkan/vk-module.h
@@ -1,9 +1,9 @@
 // vk-module.h
 #pragma once
 
-#include "../../slang.h"
+#include "slang.h"
 
-#include "../../slang-com-helper.h"
+#include "slang-com-helper.h"
 
 #if SLANG_WINDOWS_FAMILY
 #   define VK_USE_PLATFORM_WIN32_KHR 1
