+++ tools/gfx/vulkan/vk-swap-chain.h
@@ -6,7 +6,7 @@
 
 #include "../render.h"
 
-#include "../../source/core/slang-list.h"
+#include "core/slang-list.h"
 
 namespace gfx {
