+++ tools/graphics-app-framework/model.h
@@ -1,7 +1,7 @@
 // model.h
 #pragma once
 
-#include "render.h"
+#include "tools/gfx/render.h"
 #include "vector-math.h"
 
 #include <vector>
