+++ tools/graphics-app-framework/vector-math.h
