+++ tools/graphics-app-framework/windows/win-window.cpp
